requirements.txt
pandas~=2.3.3
torch~=2.9.1
tqdm~=4.67.1
sentence-transformers~=5.1.2
huggingface-hub~=0.36.0
umap-learn~=0.5.5
hdbscan~=0.8.38
scikit-learn~=1.5.0
matplotlib~=3.9.0
seaborn~=0.13.0
plotly~=5.24.0
scipy~=1.13.0
numpy~=1.26.0